LLM/requirements.txt
langchain>=0.1.0
langchain-ollama>=0.1.0
python-dotenv>=1.0.0
ollama>=0.1.0
Pillow>=10.0.0
